scraper/requirements.txt
httpx
psycopg2-binary
detoxify
torch
